requirements.txt
boto3==1.26.91
botocore==1.29.91
click==8.1.3
jmespath==1.0.1
prettytable==3.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
tqdm==4.65.0
urllib3==1.26.15
wcwidth==0.2.6
